feat(model): P15a chaos param storage narrowing (Routable<int> → Routable<int8_t>)
4 chaos fields (amount/rate/param1/param2) all have ranges 0-100
or 0-127, fitting in int8_t. Changes CurveSequence storage:

- _chaosAmount: Routable<int> (8B) → Routable<int8_t> (2B)
- _chaosRate:   Routable<int> (8B) → Routable<int8_t> (2B)
- _chaosParam1: Routable<int> (8B) → Routable<int8_t> (2B)
- _chaosParam2: Routable<int> (8B) → Routable<int8_t> (2B)

Savings: 6B each × 4 = 24B per CurveSequence × 17 patterns × 8 tracks
≈ 3,264 B .bss. Measured: .bss 118,272 → 114,760 (-3,512 B).

Accessor return types stay int for API compatibility; all getters
auto-widen int8_t → int. Setters already clamp to valid range.

P15a of CurveSequence header packing. Float fields (wavefolder/gain/
djFilter) remain as follow-up (P15b) after hardware verification.

+++ src/apps/sequencer/model/CurveSequence.h
@@ -629,12 +629,12 @@ private:
     Routable<float> _djFilter;
     float _xFade;  // Non-routable, UI-only control
 
-    Routable<int> _chaosAmount;
+    Routable<int8_t> _chaosAmount;
     ChaosAlgorithm _chaosAlgo;
     ChaosRange _chaosRange;
-    Routable<int> _chaosRate;
-    Routable<int> _chaosParam1;
-    Routable<int> _chaosParam2;
+    Routable<int8_t> _chaosRate;
+    Routable<int8_t> _chaosParam1;
+    Routable<int8_t> _chaosParam2;
 
     StepArray _steps;
